+++ lib/smartdown/api/question_page.rb
@@ -1,9 +1,4 @@
-require 'smartdown/api/multiple_choice'
-require 'smartdown/api/date_question'
-require 'smartdown/api/country_question'
-require 'smartdown/api/salary_question'
-require 'smartdown/api/text_question'
-require 'smartdown/api/postcode_question'
+require 'smartdown/model/element/question'
 
 module Smartdown
   module Api
@@ -12,19 +7,9 @@ module Smartdown
         elements.slice_before do |element|
           element.is_a? Smartdown::Model::Element::MarkdownHeading
         end.map do |question_element_group|
-          if question_element_group.find{|element| element.is_a? Smartdown::Model::Element::Question::MultipleChoice }
-            Smartdown::Api::MultipleChoice.new(question_element_group)
-          elsif question_element_group.find{ |element| element.is_a?(Smartdown::Model::Element::Question::Country) }
-            Smartdown::Api::CountryQuestion.new(question_element_group)
-          elsif question_element_group.find{|element| element.is_a? Smartdown::Model::Element::Question::Date}
-            Smartdown::Api::DateQuestion.new(question_element_group)
-          elsif question_element_group.find{|element| element.is_a? Smartdown::Model::Element::Question::Salary}
-            Smartdown::Api::SalaryQuestion.new(question_element_group)
-          elsif question_element_group.find{|element| element.is_a? Smartdown::Model::Element::Question::Text}
-            Smartdown::Api::TextQuestion.new(question_element_group)
-          elsif question_element_group.find{|element| element.is_a? Smartdown::Model::Element::Question::Postcode}
-            Smartdown::Api::PostcodeQuestion.new(question_element_group)
-          end
+          question, answer = Smartdown::Model::Element::Question.
+              create_question_answer(question_element_group)
+          question
         end.compact
       end
     end