+++ lib/smartdown/parser/node_parser.rb
@@ -4,6 +4,7 @@ require 'smartdown/parser/element/front_matter'
 require 'smartdown/parser/element/start_button'
 require 'smartdown/parser/element/multiple_choice_question'
 require 'smartdown/parser/element/date_question'
+require 'smartdown/parser/element/money_question'
 require 'smartdown/parser/element/salary_question'
 require 'smartdown/parser/element/text_question'
 require 'smartdown/parser/element/country_question'
@@ -26,6 +27,7 @@ module Smartdown
         Element::TextQuestion.new |
         Element::CountryQuestion.new |
         Element::PostcodeQuestion.new |
+        Element::MoneyQuestion.new |
         Rules.new |
         Element::StartButton.new |
         Element::NextSteps.new |
@@ -34,7 +36,7 @@ module Smartdown
       }
 
       rule(:markdown_elements) {
-        markdown_element.repeat(1, 1) >> 
+        markdown_element.repeat(1, 1) >>
         (newline.repeat(1).as(:blank_line) >> markdown_element.as(:element)).repeat
       }