+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.xx.x
+
+* money question
+
+```[money: court_fee]```
+
 ## 0.12.1
 
 Bug: ignore blank line between question headings when building question pages